+++ tests/test_pytorch/test_accelerator.py
@@ -45,6 +45,7 @@ def test_availability():
     assert HPUAccelerator.is_available()
 
 
+@pytest.mark.standalone()
 def test_all_stages(tmpdir, hpus):
     """Tests all the model stages using BoringModel on HPU."""
     model = BoringModel()
@@ -399,7 +400,8 @@ def test_hpu_parallel_reduce_op_strategy_default():
     assert strategy.reduce_op == "sum"
 
 
-@pytest.mark.skip(reason="TBD : Fix pytest issues")
+@pytest.mark.standalone()
+@pytest.mark.skip(reason="fix ddp pytest issue")
 @pytest.mark.skipif(HPUAccelerator.auto_device_count() < 2, reason="Test requires multiple HPU devices")
 @pytest.mark.parametrize(
     ("reduce_op", "expectation"),
@@ -460,7 +462,8 @@ def test_reduce_op_strategy(tmpdir, hpus, reduce_op, expectation):
         trainer.fit(_model)
 
 
-@pytest.mark.skip(reason="TBD : Fix pytest issues")
+@pytest.mark.standalone()
+@pytest.mark.skip(reason="fix ddp pytest issue")
 @pytest.mark.skipif(HPUAccelerator.auto_device_count() < 2, reason="Test requires multiple HPU devices")
 @pytest.mark.parametrize(
     ("reduce_op", "logged_value_epoch", "logged_value_step"),